// apps/web/src/types/Node.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { NodeType } from "./NodeType";

export type Node = { id: string, category: string, node_type: NodeType, text: string, semantic_id: string | null, display_category: string | null, position_x: number | null, position_y: number | null, is_active: boolean, created_at: string, updated_at: string, };